Updated project to include full functionality, minor changes to styles.css and index.html.
Changed title on index.html to be more descriptive of the actual project. Adjusted a rule declaration in styles.css so that it properly points to the crossOutButton class of li elements.

Updated scripts.js to include functionality required from Task 1.9, all written in jQuery. Users can now add, delete, cross out, and reorder list elements of this to-do list app.